Finished the basic Etch-A-Sketch
I said basic because I couldn't make it do what I wanted. I couldn't
get the divs to color when I dragged the mouse over them. That was the
only thing that kept me from moving on.

But I have the basic project working, so I'll call this a success.